Added absolute file imports and moved code to /src folder
