+++ web-app/src/stories/common/Button.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Button } from "../../components/common/Button";
+import {Button} from "@/components/common/Button";
 
 export default {
     title: 'Common/Button',
