+++ web-app/src/stories/examples/ExampleComponent.stories.js
