+++ web-app/src/stories/layout/Header.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Header} from "../../components/layout/Header";
+import {Header} from "@/components/layout/Header";
 
 export default {
     title: 'Layout/Header',
